Handle 404 on download

+++ prettier-maven-plugin/src/main/java/com/hubspot/maven/plugins/prettier/internal/NodeDownloader.java
@@ -2,32 +2,17 @@ package com.hubspot.maven.plugins.prettier.internal;
 
 import java.io.IOException;
 import java.io.InputStream;
-import java.nio.file.DirectoryNotEmptyException;
-import java.nio.file.FileAlreadyExistsException;
-import java.nio.file.FileSystemException;
 import java.nio.file.Files;
 import java.nio.file.Path;
 import java.nio.file.Paths;
 import java.nio.file.StandardCopyOption;
-import java.nio.file.StandardOpenOption;
-import java.nio.file.attribute.FileAttribute;
-import java.nio.file.attribute.PosixFilePermission;
-import java.nio.file.attribute.PosixFilePermissions;
-import java.util.Arrays;
-import java.util.Set;
-import java.util.zip.ZipEntry;
-import java.util.zip.ZipInputStream;
-
-import org.apache.commons.compress.archivers.tar.TarArchiveEntry;
-import org.apache.commons.compress.archivers.tar.TarArchiveInputStream;
-import org.apache.commons.compress.compressors.gzip.GzipCompressorInputStream;
-import org.apache.maven.plugin.MojoExecutionException;
-import org.apache.maven.plugin.logging.Log;
-
+import java.util.Optional;
 import okhttp3.OkHttpClient;
 import okhttp3.Request;
 import okhttp3.Response;
 import okhttp3.ResponseBody;
+import org.apache.maven.plugin.MojoExecutionException;
+import org.apache.maven.plugin.logging.Log;
 
 public class NodeDownloader {
   private static final OkHttpClient HTTP_CLIENT = new OkHttpClient();
@@ -43,7 +28,10 @@ public class NodeDownloader {
   public NodeInstall download(String version) throws MojoExecutionException, IOException {
     OperatingSystemFamily os = OperatingSystemFamily.current();
     log.debug("Determined os: " + os);
+    return download(version, os);
+  }
 
+  public NodeInstall download(String version, OperatingSystemFamily os) throws MojoExecutionException, IOException {
     Path targetDirectory = installDirectory.resolve(os.getNodeDirectoryName(version));
     if (Files.exists(targetDirectory)) {
       log.debug("Reusing cached node at: " + targetDirectory);
@@ -51,7 +39,17 @@ public class NodeDownloader {
       String downloadUrl = os.getNodeDownloadUrl(version);
       log.debug("Downloading node from url: " + downloadUrl);
 
-      Path nodeArchive = downloadToTmpFile(downloadUrl);
+      Optional<Path> maybeNodeArchive = downloadToTmpFile(downloadUrl);
+      if (!maybeNodeArchive.isPresent()) {
+        Optional<OperatingSystemFamily> fallback = os.getFallback();
+        if (fallback.isPresent()) {
+          return download(version, fallback.get());
+        } else {
+          throw new MojoExecutionException("Got 404 when trying to download node from: " + downloadUrl);
+        }
+      }
+
+      Path nodeArchive = maybeNodeArchive.get();
       log.debug("Downloaded node to: " + nodeArchive);
 
       Path tmpDir = os.extractToTmpDir(installDirectory, nodeArchive);
@@ -73,16 +71,25 @@ public class NodeDownloader {
     return os.toNodeInstall(targetDirectory);
   }
 
-  private Path downloadToTmpFile(String downloadUrl) throws IOException {
+  private Optional<Path> downloadToTmpFile(String downloadUrl) throws MojoExecutionException, IOException {
     Request request = new Request.Builder().url(downloadUrl).build();
 
     try (Response response = HTTP_CLIENT.newCall(request).execute();
          InputStream responseStream = responseStream(response)) {
-      Path tempFile = Files.createTempFile(installDirectory, "node-", ".tmp");
-      Files.copy(responseStream, tempFile, StandardCopyOption.REPLACE_EXISTING);
-      tempFile.toFile().deleteOnExit();
+      if (response.code() == 404) {
+        log.debug("Got 404 when trying to download node from: " + downloadUrl);
+        return Optional.empty();
+      } else if (!response.isSuccessful()) {
+        throw new MojoExecutionException(
+            "Got response code " + response.code() + " when trying to download node from: " + downloadUrl
+        );
+      } else {
+        Path tempFile = Files.createTempFile(installDirectory, "node-", ".tmp");
+        Files.copy(responseStream, tempFile, StandardCopyOption.REPLACE_EXISTING);
+        tempFile.toFile().deleteOnExit();
 
-      return tempFile;
+        return Optional.of(tempFile);
+      }
     }
   }
